Create Weeks class and derive WeekRange from Weeks
Deserialize JSON into Weeks or WeekRange depending on the fields it has
