+++ app/src/main/java/com/ashuh/nusmoduleplanner/api/nusmods/NusModsApiInterfaceBuilder.java
@@ -2,11 +2,11 @@ package com.ashuh.nusmoduleplanner.api.nusmods;
 
 import com.ashuh.nusmoduleplanner.api.nusmods.deserializer.PrereqTreeDeserializer;
 import com.ashuh.nusmoduleplanner.api.nusmods.deserializer.SemesterTypeDeserializer;
-import com.ashuh.nusmoduleplanner.api.nusmods.deserializer.WeekRangeDeserializer;
+import com.ashuh.nusmoduleplanner.api.nusmods.deserializer.WeeksDeserializer;
 import com.ashuh.nusmoduleplanner.api.nusmods.deserializer.WorkloadDeserializer;
 import com.ashuh.nusmoduleplanner.data.model.nusmods.module.prereqtree.PrereqTree;
 import com.ashuh.nusmoduleplanner.data.model.nusmods.module.semesterdatum.SemesterType;
-import com.ashuh.nusmoduleplanner.data.model.nusmods.module.semesterdatum.lesson.weekrange.WeekRange;
+import com.ashuh.nusmoduleplanner.data.model.nusmods.module.semesterdatum.lesson.weeks.Weeks;
 import com.ashuh.nusmoduleplanner.data.model.nusmods.module.workload.Workload;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -21,7 +21,7 @@ public class NusModsApiInterfaceBuilder {
         GsonBuilder gsonBuilder = new GsonBuilder()
                 .registerTypeAdapter(PrereqTree.class, new PrereqTreeDeserializer())
                 .registerTypeAdapter(SemesterType.class, new SemesterTypeDeserializer())
-                .registerTypeAdapter(WeekRange.class, new WeekRangeDeserializer())
+                .registerTypeAdapter(Weeks.class, new WeeksDeserializer())
                 .registerTypeAdapter(Workload.class, new WorkloadDeserializer());
         Gson gson = gsonBuilder.create();
